pyproject.toml
[tool.poetry]
name = "airbnb-clone-backend"
version = "0.1.0"
description = ""
authors = ["sds7629 <sds7629@naver.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "airbnb_clone_backend"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
pillow = "^9.5.0"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
pyjwt = "^2.8.0"
django-environ = "^0.10.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
